Fix Linux compilation errors with bin_debug_severe in MPI mode
- Running the command `make CPPFLAGS=-DSWMPI bin_debug_severe` on Linux/HPC environment results in the following errors
	* src/SW_MPI.c: In function ‘fillDesignationIO’:
	  src/SW_MPI.c:751:22: error: operand of ‘?:’ changes signedness from ‘int’ to ‘size_t’ {aka ‘long unsigned int’} due to unsignedness of other operand [-Werror=sign-compare]
          751 |                      desig->nCompProcs :
	* src/SW_MPI.c: In function ‘alloc_IO_info’:
	  src/SW_MPI.c:2495:37: error: comparison of integer expressions of different signedness: ‘size_t’ {aka ‘long unsigned int’} and ‘int’ [-Werror=sign-compare]
          2495 |     for (allocIndex = 0; allocIndex < num1DArr; allocIndex++) {

- Cast "nCompProcs" to "size_t"
- Convert type of "num1DArr" to "size_t"

+++ src/SW_MPI.c
@@ -748,7 +748,7 @@ static void fillDesignationIO(
         desig->nSuids = *leftSuids;
     } else if (*leftOverSuids > 0) {
         nSuids = (*leftOverSuids >= (IntU) desig->nCompProcs) ?
-                     desig->nCompProcs :
+                     (size_t) desig->nCompProcs :
                      *leftOverSuids;
 
         desig->nSuids += nSuids;
@@ -2442,7 +2442,7 @@ static void alloc_IO_info(
 ) {
     size_t suid;
     size_t allocIndex;
-    const int num1DArr = 1;
+    const size_t num1DArr = 1;
     int inKey;
     size_t numVals;
     size_t nDomInfoElem = maxSuidsInWrite * N_ITER_BEFORE_OUT;
